requirements.txt
flask
pydantic
python-dotenv
pymupdf
chromadb
sentence-transformers
tiktoken
openai